refactor: 콘솔 관련 코드를 console/ 디렉토리로 응집
