+++ src/js/step1-index.js
@@ -3,9 +3,10 @@
  * 브라우저 환경에서 사용하는 css 파일 등을 불러올 경우 정상적으로 빌드할 수 없습니다.
  */
 
-import ConsoleView from './UI/console/View.js';
-import { runConsoleStateMachine } from './controller/consoleStateMachine.js';
+import ConsoleView from './console/View.js';
+import LottoGame from './controller/LottoGame.js';
+import { runConsoleLotto } from './console/controller.js';
 
 const view = new ConsoleView();
 
-runConsoleStateMachine(view, { haltOnError: true });
+runConsoleLotto(view, new LottoGame());
