+++ src/js/step2-index.js
@@ -3,10 +3,11 @@
  * 브라우저 환경에서 사용하는 css 파일 등을 불러올 경우 정상적으로 빌드할 수 없습니다.
  */
 
-import { runConsoleStateMachine } from './controller/consoleStateMachine.js';
-import withRetry from './controller/withRetry.js';
-import ConsoleView from './UI/console/View.js';
+import ConsoleView from './console/View.js';
+import LottoGame from './controller/LottoGame.js';
+import { runConsoleLotto } from './console/controller.js';
+import withRetry from './console/withRetry.js';
 
 const view = new ConsoleView();
 
-withRetry(runConsoleStateMachine, view, { haltOnError: false });
+withRetry(() => runConsoleLotto(view, new LottoGame()), view);
